Rename RDM to RSA.
