+++ src/analysis/rsa/bandpass/mean_rdms.py
@@ -12,7 +12,7 @@ from scipy.spatial.distance import squareform, pdist
 current_dir = pathlib.Path(os.path.abspath(__file__)).parent
 sys.path.append(os.path.join(str(current_dir), "../../../../"))
 
-from src.analysis.rsa.rdm import alexnet_layers
+from src.analysis.rsa.rsa import alexnet_layers
 from src.analysis.rsa.activations import load_activations
